refactor geoparser structure to consolidate module interfaces and implement lazy-loading for modules
