+++ geoparser/__init__.py
@@ -1,2 +1,3 @@
 from geoparser.geoparser import Geoparser
+from geoparser.geoparserv2 import GeoparserV2
 from geoparser.trainer import GeoparserTrainer
