+++ geoparser/geoparserv2/__init__.py
@@ -1,6 +1,1 @@
 from geoparser.geoparserv2.geoparserv2 import GeoparserV2
-from geoparser.geoparserv2.modules.interfaces import (
-    BaseModule,
-    RecognitionModule,
-    ResolutionModule,
-)
