--- geoparser/geoparserv2/modules/__init__.py
@@ -1,5 +1,0 @@
-from geoparser.geoparserv2.modules.interfaces import (
-    BaseModule,
-    RecognitionModule,
-    ResolutionModule,
-)
