+++ tests/modules/test_interfaces.py
@@ -3,11 +3,7 @@ import json
 
 import pytest
 
-from geoparser.geoparserv2.modules.interfaces import (
-    BaseModule,
-    RecognitionModule,
-    ResolutionModule,
-)
+from geoparser.modules.interfaces import BaseModule, RecognitionModule, ResolutionModule
 
 
 def test_base_module_init():
